create validator.ts
