+++ src/$public/$public.ts
@@ -1,7 +1,9 @@
 import { config } from './config';
 import { errors } from './errors';
+import { validator } from './validator';
 
 export const $public = {
   errors,
   config,
+  validator,
 };
